20210716 Create .exe building and testing for PR pipeline
