+++ azure-pipelines.yml
@@ -488,3 +488,38 @@ jobs:
           scanType: "Register"
           verbosity: "Verbose"
           alertWarningLevel: "High"
+
+  - job: exetest
+    displayName: .exe Test Windows
+    pool:
+      vmImage: "vs2017-win2016"
+    strategy:
+      matrix:
+        node_8_x:
+          node_version: 8.x
+        node_10_x:
+          node_version: 10.x
+        node_12_x:
+          node_version: 12.x
+    steps:
+      - task: NodeTool@0
+        inputs:
+          versionSpec: "$(node_version)"
+        displayName: "Install Node.js"
+
+      - script: |
+          npm ci
+        workingDirectory: "./"
+        displayName: "npm ci"
+
+      - script: |
+          npm run build:exe
+        workingDirectory: "./"
+        displayName: "npm run build:exe"
+        env: {}
+
+      - script: |
+          npm run test:exe
+        workingDirectory: "./"
+        displayName: "npm run test:exe"
+        env: {}
